+++ src/components/admin/CreateSessionPage.js
@@ -1,49 +1,14 @@
 import React, { useState, useEffect } from "react";
-import { MainButton } from "../common/Button";
 import { Container, InputContainer } from "../common/Container";
-import { StyledInput } from "../common/Input";
-import { useCreateSession } from "../../viewModel/adminHook";
 import { checkAttendStart } from "../../utils/authentication";
-import Logo from "../common/Logo";
 import { Header } from "../common/Header";
 import { useNavigate } from "react-router-dom";
 import { checkAdminState } from "../../utils/authentication";
 import {Gap} from "../common/Gap";
+import {CreateSessionForm} from "./CreateSessionForm";
 
 
 
-const CreateSessionForm = ({ navigate }) => {
-  const onSuccess = () => {
-    navigate("/sessions");
-  };
-
-  const {
-    sessionName,
-    date,
-    onChangeSessionName,
-    onChangeDate,
-    onPressCreateSession,
-  } = useCreateSession(onSuccess);
-
-  return (
-    <InputContainer>
-      <StyledInput
-        placeholder="세션 이름"
-        keyboardType="default"
-        value={sessionName}
-        onChangeText={onChangeSessionName}
-        maxLength={50}
-      />
-      <StyledInput
-        type="date"
-        placeholder="세션 날짜(YYYY-MM-DD)"
-        value={date}
-        onChangeText={onChangeDate}
-      />
-      <MainButton content="세션 생성" onPress={onPressCreateSession} />
-    </InputContainer>
-  );
-};
 
 const CreateSessionPage = () => {
   const [isStart, setIsStart] = useState(true);
