+++ src/components/admin/UpdateUser.js
@@ -1,67 +1,18 @@
-import React, { useState, useEffect } from "react";
-import { useNavigate, useLocation } from "react-router-dom";
-import styled from "styled-components";
-import { api } from "../../utils/api";
-import bcrypt from "bcryptjs";
+import React from "react";
+import {  useLocation } from "react-router-dom";
 import { Container, InputContainer } from "../../components/common/Container";
 import { StyledInput } from "../../components/common/Input";
 import { MainButton } from "../../components/common/Button";
-import Logo from "../../components/common/Logo";
 import { Header } from "../../components/common/Header";
-import { checkAdminState } from "../../utils/authentication";
 import {Gap} from "../common/Gap";
+import { useUpdateUser } from "../../viewModel/adminHook";
+
 
 const UpdateUser = () => {
-  const navigate = useNavigate();
   const location = useLocation();
   const { userId } = location.state || {};
-  const [username, setUsername] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  useEffect(() => {
-    if (!userId) {
-      console.error("User ID not provided");
-      return;
-    }
-
-    const fetchUser = async () => {
-      try {
-        const response = await api.get(`/user/users/${userId}`);
-        const user = response.data;
-        setUsername(user.username);
-        setEmail(user.email);
-        setPassword("");
-      } catch (error) {
-        console.error("Failed to fetch user:", error);
-      }
-    };
-
-    fetchUser();
-    checkAdminState(navigate);
-  }, [userId]);
-
-  const handleUpdateUser = async () => {
-    try {
-      let hashedPassword = password;
-
-      if (password) {
-        const salt = await bcrypt.genSalt(10);
-        hashedPassword = await bcrypt.hash(password, salt);
-      }
-
-      await api.put(`/user/users/${userId}`, {
-        username,
-        email,
-        password: hashedPassword,
-      });
-      alert("유저 정보가 변경되었습니다.");
-      navigate("/users");
-    } catch (error) {
-      console.error("Failed to update user:", error);
-      alert("유저 정보 변경에 실패했습니다. 다시 시도해주세요.");
-    }
-  };
+  const { username, setUsername, email, setEmail, password, setPassword, handleUpdateUser } = useUpdateUser(userId);
+  
 
   return (
     <Container>
