+++ src/components/admin/UserList.js
@@ -8,6 +8,7 @@ import { StyledText } from "../common/Text";
 import { checkAdminState } from "../../utils/authentication";
 import { Container, InputContainer, TwoButtonContainer } from "../common/Container";
 import useListDataStore from "../../states/listDataStore";
+import { useUserList } from "../../viewModel/adminHook";
 
 const UserItem = styled.div`
   display: flex;
@@ -41,27 +42,7 @@ const UpdateButton = styled.button`
 
 const UserList = () => {
   const navigate = useNavigate();
-  const { data: users, loading, error, setData, setLoading, setError } = useListDataStore();
-
-  useEffect(() => {
-    const fetchUsers = async () => {
-      setLoading(true);
-      try {
-        const response = await api.get("/user/users");
-        const filtered_user = response.data.filter(
-          (user) => user.isAdmin === false
-        );
-        setData(filtered_user);
-      } catch (err) {
-        setError(err.message);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchUsers();
-    checkAdminState(navigate);
-  }, [navigate, setData, setLoading, setError]);
+  const { users, loading, error } = useUserList();
 
   if (loading) return <Container>Loading...</Container>;
   if (error) return <Container>Error: {error}</Container>;
